Accept a `rusqlite::Transaction` in Tolstoy's sync.
This is part of a larger shift towards atomically updating `Conn`.
The division of labour is (and must be) similar to `db` and `Conn` in
a `transact`: `transact` coordinates `db` to do low-level work and
processes progress reports, and then commits the SQL transaction and
atomically advances `Conn`'s metadata.  In a parallel fashion, `sync`
will coordinate `tolstoy` to do low-level work before committing the
SQL transaction and atomically advancing `Conn`'s metadata.

Co-authored-by: Nick Alexander <nalexander@mozilla.com>